Factor out coercion.rs
Coercion logics are shared among the interpreter and the parser, so it
should be in a separate module.
