+++ parser/src/lib.rs
@@ -6,6 +6,7 @@ macro_rules! dbg_println {
 }
 
 mod bytecode;
+mod coercion;
 mod compiler;
 mod format_ast;
 mod interpreter;
@@ -20,9 +21,10 @@ mod vm;
 pub use self::bytecode::{
     Bytecode, DebugInfo, FnBytecode, FunctionInfo, Instruction, LineInfo, OpCode,
 };
+pub use self::coercion::coerce_type;
 pub use self::compiler::*;
 pub use self::format_ast::format_stmts;
-pub use self::interpreter::{coerce_type, run, EvalContext, EvalError, FuncDef};
+pub use self::interpreter::{run, EvalContext, EvalError, FuncDef};
 pub use self::parser::{span_source as source, ArgDecl, ReadError, Span};
 pub use self::type_checker::{type_check, TypeCheckContext};
 pub use self::type_decl::TypeDecl;
